Feat: 미팅 관련 클래스, 미팅 예약하기 기능 구현
Former-commit-id: bccf643fd90f57dfb20a070853709a4c7a0bd018
Former-commit-id: fdb19c9da5e676c5b136e2d9ea1632bf8591a5b2

+++ backend/src/main/java/com/moi/anitime/model/repo/AnimalRepo.java
@@ -10,7 +10,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface AnimalRepo extends JpaRepository<Animal, Integer> {
+public interface AnimalRepo extends JpaRepository<Animal, Long> {
 
     @Query(value = "SELECT A.*,IF(B.bookmarkNo IS NOT NULL,'Y','N') AS isBookMarked FROM Animal AS A " +
             "LEFT JOIN Bookmark AS B ON A.desertionNo=B.desertionNo AND B.generalNo=:generalNo"+
